lib: check for UINPUT_ properties to allow for emulated devices
Emulating devices is important for debugging when the hardware isn't
present. udev doesn't apply the same settings to emulated uinput devices as
it does to real devices, so we need to guess.

To allow some flexibility, use an UINPUT_ namespace and check that. Custom
udev rules may then set the UINPUT_SUBSYSTEM for an emulated device.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>