Generate extra udev rules to assing fields to uinput devices
uinput devices aren't processed like other devices in udev, they don't get
the right hierarchy and not all the properties are assigned.
uinput devices are useful for debugging where the hardware isn't available
though.

Add a --with-uinput-rules flag to the generate-udev-rules helper tool to
write a rule for each device in the database to apply extra properties if
that device were a virtual device.

New properties assigned:
  ENV{UINPUT_DEVICE}="1"
  ENV{UINPUT_SUBSYSTEM}="<subsystem>"

(udev doesn't allow overriding the subystem)

Other properties assigned:
  ENV{ID_VENDOR_ID}
  ENV{ID_MODEL_ID}

Note that these properties will be available on the parent only, there is
nothing on the child we can match on.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>