test: remove special-case for DTI-520
We don't need to special-case the device, this device may be used with a pen
that includes an eraser. Removing the special case means we expect all
currently known devices to have at least one stylus, the code from
4a7a7f16564869b3af6701aa943e5c88927f451b tests those styli for additional
eraser.

Quote Ping Cheng, 17 May 2012,
linuxwacom-devel, Thread: Add support for the DTI-520
"DTI-520 is a protocol 4 (V4) device. It does not support serial number
and tool IDs. [...] This is the same as Graphires. Even if the pen included
in the box does not have eraser, user may use a pen with eraser."

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Bastien Nocera <hadess@hadess.net>