rename: Bamboo Pen & Touch data file
Rename bamboo-pen-and-touch.tablet to bamboo-2fg-4x5.tablet.  It is a first generation BambooPT.  The tablet drawing surface is larger than the pen's active area on 3 sides.  The active area is a rectangle delineated on 3 sides by a white line about 1 cm in from the bezel edge and on the fourth side by the pad buttons.  The touch active area is inside and smaller than the pen's.  This arrangement means your finger does not have to contact a button edge.

Signed-off-by: David Foley <favux.is@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>