data: update the database entry with status LED
Adds a new field "StatusLEDs" in the feature group
to list the status LEDs present on the device in the
same order as they appear in the sysfs control entries.

For example, using "StatusLEDs=Ring2;Ring" means that
the 2nd ring mode switch buttons (ie listed as "Ring2")
would control the LED available at status_led0_select
(since it's in first position in the "StatusLEDs" list)
and the 1st ring mode switch buttons (ie listed as "Ring")
would control the LED available at status_led1_select
(since it's in second position in the "StatusLEDs" list).

More status LED not necessarily correlated with mode switch
button may be added in the future, as needs arise.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>
Reviewed-by: Jason Gerecke <killertofu@gmail.com>