re-institute AM_MAINTAINER_MODE (default enabled)
See Ron's comments discussion on:
http://old.nabble.com/-PATCH-libwacom--Drop-AM_MAINTAINER_MODE-td34561358.html

Summary:
In this workflow all generated files are committed to git to have a known
working set. On checkout, git does not restore the original timestamps,
so with maintainer-mode enabled, automake will regenerate the files - which
is to be avoided since we already have the known working set.

AM_MAINTAINER_MODE([enable]) gives us the same behaviour as now, but will
allow Ron to disable the regeneration of the files.

Requested-by: Ron <ron@debian.org>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>