data: add leaders to layouts
To match the buttons with their corresponding labels, the
layout SVG must also provide a leader line for each label
in the form of a line that links each button and its
label.

Improves the layout SVG for the Cintiq 12WX and Intuos 5 M
by adding details and rounded buttons to match the look of
the real device.

Buttons, labels and indicators are now grouped in the SVG
layout for better clarity of the SVG.

Replace use of ambiguous "Up" and "Down" in ring
controls with "CW" (Clockwise) and "CCW" (Counter
Clockwise) which are more appropriate for circular
motion (note that "CCW" is used in place of "Up"
and "CW" in place of "Down").

Add possible use of "fake" buttons to materialize
the type of motion expected in touch strips/rings
(optional, not mandatory as not present on the
real device).

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>