libwacom: [PATCH] lib: use integration flags from device info
Resend, sorry the commit message was the opposite of what I meant...

>From 378edbc1b5caec1e27164d33dd4135a38b32f003 Mon Sep 17 00:00:00 2001
From: Olivier Fourdan <ofourdan@redhat.com>
Date: Mon, 3 Dec 2012 10:25:11 +0100
Subject: [PATCH] lib: use integration flags from device info

if and only if the entry in the database did not set
a value.

The flags set in the database entry should not be changed
based on kernel flags unless the database did not set the
integration flags.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>