test: check number of switches against number of modes
Most tablets have one mode-switch button and multiple
modes, but some like the Cintiq 24HD have dedicated mode-
switch buttons.

In this case, the number of modes must match the number of
switches otherwise client applications have no way to tell
if a button is used as a mode-switch button to cycle through
all the modes or a mode-switch button to select a given mode.

Add this in the tablet-validity check, ie make sure that
if we have more than one mode-switch button, then the
number of modes matches the number of mode-switch buttons.

Also checks that if more than one mode is declared, at least
one mode-switch button is found.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>