Rely on the PRODUCT attribute to fetch the bus, vid & pid informations
PRODUCT attribute is set correctly for the input device, whether the
device is USB or Bluetooth. We can use a common path for those two.
uinput devices have their PRODUCT attribute set according to the
registered device, so they work too.

This change allows libwacom to properly detect the actual PID of the
tablet connected through the wireless receiver when the following kernel
patch (or its sucessor) will be merged in 3.18:
https://patchwork.kernel.org/patch/4889811/

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Reviewed-by: Bastien Nocera <hadess@hadess.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>