Allow matching on the name too
To support generic tablets, we need to extend the matching pattern of
libwacom. The Huion tablets reuse the same VID:PID accross all of the
products. To be able to differentiate between the models, we can somewhat
rely on the name of the product. It doesn't seem very reliable (for
instance, the H610 Pro and the W58 seems to share the same name), but it
should allow a good amount of them to be added to the DB.

The current usb:VID:PID pattern is still working, we add a new pattern
usb:VID:PID:NAME.

For devices with their name in the matching pattern, the function
libwacom_new_from_usbid() does not work. We encourage users to use
libwacom_new_from_path() instead. (g-c-c and g-s-d already use that, and
libinput just switched to it).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>