tools: match on ATTRS, not ENV to apply to the uinput event node directly
ENV{PRODUCT} is set only on the /sys/class/input/inputXX node, not on the
event node itself. Thus, we never set the properties on the device node
itself.

Match on ATTRS{id/vendor} and {id/product} instead, this way we can apply the
various ID_INPUT_ tags directly on the node.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>