ek-remote: fix the buttons declaration
The ASCII art had a glitch.

The buttons are not left/right. They are all grouped into the same logical
group.

Note that g-s-d now needs to be updated to not count on the button flags
to place the label, but actually rely on the SVG position which is correct.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pingc@wacom.com>