data: add makefile hook to verify all files are in git
Use a trick to verify git is there. configure.ac falls back to /usr/bin/false
if git is missing from the path, so any git command will fail. Since we check
whether we're in a git directory first (to handle tarballs) this will fail
with /usr/bin/false and that code-path is not met.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Jason Gerecke <jason.gerecke@wacom.com>