tools: Add udev rules quirk for Wireless Accessory Kit
Although the kernel modifies the PID of devices connected to the system
through Wacom's Wireless Accessory Kit to match that of the actual
tablet, it does not look like udev is aware of this change and instead
matches on the VID/PID of the kit's dongle. To ensure devices connected
this way are properly tagged, add general-purpose matches for the kit.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>