layouts: Center the EKR layout on the document
The gnome OSD uses the document size in order to display the SVG,
this is usually desirable as pad buttons must be placed relative
to their real location in the tablet.

In the EKR case, the layout sits somewhere in between in the
document, not even centered horizontally or vertically, so applying
the same sizing logic that's usually desirable for integrated pads
comes out with the EKR layout sitting in an equally seemingly random
place in the screen.

Changing the document size to fit the layout better results in
the OSD layout being nicely centered, and using the maximum space
available.

Signed-off-by: Carlos Garnacho <carlosg@gnome.org>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>