test: Require reversible tablets to have left/right buttons
The "reversible" flag is supposed to indicate if a device can be rotated
180 degrees for left-handed use. A number of tablets currently set this
flag even though they only have buttons along the top, which makes no
sense -- rotating the tablet (though possible) gains a left-handed user
absolutely nothing.

This updates the device tests to require that reversible tablets have
buttons on the left or right. The no-button case is specifically ignored
to allow several tablets in a family of reversible tablets to continue
to pass.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>