Add libwacom_get_paired_device()
On some devices we have a different PID for the touch device which makes
grouping the two devices harder in userspace. Provide a call to fetch those.

The call is identical to the DeviceMatch approach but only one ID is allowed
for the PairedID. It would get more difficult if we have an NxM match to
paired ID, in that case it's likely best to have separate data files.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Aaron Armstrong Skomra <aaron.skomra@wacom.com>