data: enable modes on the ExpressKey Remote
While it's true modes are read-only and it doesn't change a bit on X.org,
under wayland, we need to have the number of modes and the toggle key.

The first change in X.org is that now g-s-d calls gsd-wacom-led-helper
to try to set the LED, but this just fails and no one cares actually.

The second change is that now the OSD shows modes and might be out of
sync with the actual mode in use. It's not a big issue as right now
g-c-c doesn't show anything for the EKR so users who are actually using
the OSD played with dconf and are not beginners. Given that g-c-c needs
fixes for the EKR, we can also make sure g-s-d updates the mode according
to what the hardware provides.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Acked-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>