data: Add data and SVG files for MobileStudio Pro
Note that the MobileStudio Pro introduces a new "Pro Pen" which supports
8K pressure levels. The ID of this pen is identical to that of a (less
featureful) pen used by the Intuos2, so we have to slightly modify the
existing entry. Ideally, we would have some way to define multiple pens
which share the same IDs (which apparently, according to the technical
experts, are not guaranteed unique outside of a hardware generation).
Such a change will require serious design and planning that are outside
of the immediate scope of this patch.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>