Add new 'EvdevCodes' key to the Buttons section
The button number to evdev code mapping is non-predictable on some tablets and
needs to be hardcoded in the device files.

Some heuristics is employed here so we don't have to add it to all files, only
the ones that fall out of the heuristics for a specific class.

The Huion 610 Pro currently warns about unknown mappings. This requires
someone with that device to fix it.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>