Don't double-free the name during matching
The name is freed a few lines down as we fall through to the 'bail:' label.
The only exception is when we actually return the device, that's where we need
to free it manually.

This is just to shut up coverity, this code can only be triggered by supplying
an invalid WacomFallbackFlag.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>