Add a ModelName property to all devices
Some of the names are overlapping or at least hard to identify, it's useful to
be able to use the Wacom-assigned names (e.g. PTH-660) together with the
marketing name. Doubly so in the Intuos vs Bamboo-is-now-Intuos series.

Where there isn't an obvious model name it is NULL. This includes the various
ISDV4 devices that could use that as model name but in those cases that name
is encoded already in the human-readable name anyway.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>