Reduce the number of dprintfs in the button printing
No functional changes, the output is still the same. This is prep work for
a follow-up patch.

Instead of a full check and alloca of everything in the perfect size, I opted
for large enough buffers with a few asserts instead. These should never be
triggered on correct .device files anyway, so it's good enough.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>