Print lines with empty data as comment
Affects the StatusLEDs= line and the various button (Right=, Left=, ...)
lines.

We don't add these in the real data files if we don't have entries for them.
Let's print them as comments here (instead of leaving them out) to make it
obvious that a) they have been processed and b) there's really nothing to see
here.

The IntegratedIn= line is added even where empty, and others (Styli and
EvdevCodes) are filled in even where empty in the original file. Let's
leave those as-is for now.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>