Add support for the meson build system
Only significant change in the installed directory structure: there is no
static library built anymore.

General usage:
 $ meson builddir
 $ ninja -C builddir
 $ ninja -C builddir test
 $ ninja -C builddir install

To run scan-build:
 $ ninja -C builddir scan-build

To run ubsan:
 $ meson configure builddir -Db_sanitize=undefined
 $ ninja -C builddir test

To run asan:
 $ meson configure builddir -Db_sanitize=address
 $ ninja -C builddir test

meson also supports test setups where the test suite can be wrapped by another
exectable, in this case valgrind. This is hooked up for some of the tests
(using test suites) so we can skip the ones we don't want to run in valgrind.
Best invocation is now:

 $ cd builddir
 $ meson test --setup=valgrind --suite=valgrind

glib always leaks the types, so we need the valgrind suppressions file to
avoid false positives.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>