Switch from implicit symbol exports to explicit ones
Don't just export anything starting with "libwacom_", instead mark the
functions we do want to export as visibility default and enable a default
symbol visbility as hidden.

We just assume that -fvisibility=hiddeen works, there are compilers that need
different handling, but lets add them as needed..

libwacom_error_set_message() wasn't supposed to be ABI as evidenced by the
comment, but it got scooped up by the regex. So ABI it remains, we should
probably make it a noop one day.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>