Return an error if the wrong fallback flags have been given
This is one of the few functions where we do GError handling, so we can't just
throw a fit and return NULL. We have to return a proper error code because we
expect callers to potentially check the error on failure.

So let's introduce a new error code to denote general caller bugs.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>