Deprecate libwacom_error_set()
This function wasn't intended to become API. It's not documented either and
not available in a public header file, so the only way to call it was to dlsym
it or to declare it locally - neither of which makes sense. Let's remove most
of its functionality so it cannot be abused.

The alternative is to bump the soname and remove it altogether. No-one would
likely notice but it will require rebuilding everything.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>