Deprecate libwacom_match_destroy(), libwacom_update_match(), libwacom_match_new()
These functions weren't intended to become API. They arenot documented either
and not available in a public header file, so the only way to call any of them
was to dlsym it or to declare it locally - neither of which makes sense. Let's
remove most of their functionality so they cannot be abused.

The alternative is to bump the soname and remove it altogether. No-one would
likely notice but it will require rebuilding everything.

As a side-effect: we can now get rid of libwacom_copy_match() too now. Matches
are never modified once loaded, so we can just ref them everywhere we need
them.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>