Alias the accidentally exposed ABI into different functions
We rename the previously exported symbols to something else that doesn't
pollute our namespace. Then we use an assembly command to map the symbol into
the current ABI version. Once we change that one, the symbols will disappear
from the new version, creating a linker error. Sicne we don't expect anyone to
use them anyway, life will just go on normally for everyone.

This also clears up the actual symbols for use in our code if we need to.

For information on the asm command:
https://gcc.gnu.org/wiki/SymbolVersioning
https://akkadia.org/drepper/symbol-versioning
https://www.akkadia.org/drepper/dsohowto.pdf (p39)

A special "trick" is used here to hide the ABI from new versions:
Usually when defining multiple versioned symbols, one would define one as the
default one with @@
    .symver _foo1,foo@VERSION1
    .symver _foo2,foo@@VERSION2 <-- default one

By leaving out the default one, ld doesn't know which one to link to and
fails with an unresolved symbol. rtld however can still figure it out, so
anything compiled will continue to work. This way we can make a symbol
disappear from the library for new builds but have old builds continue to
work with the new version.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>