test: add a test for our deprecated symbols
This is the most basic test, dlopen the library and check for the sym.
It's a fairly useless test too right now since this test cannot fail as long
as our linker setup works. If it fails, everything else will fail too.

But if in the future we remove those symbols, the test will fail so that'll
tell us something - whatever that something is when we want to hear it.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>