Fix parsing of match strings when using musl
Several tests fail when running on a system with the musl libc library
with the error "Duplicate match of 'usb:256c:006e' on device 'Huion H420'".
The root cause appears to be musl's handling of '%63c' in our format string.
Specifically, unlike glibc which interprets this to mean "read up to 63
characters", musl interprets it as "read exactly 63 characters".

This commit tweaks our format string to extract the first three required
fields and probe (with ':%n') to see if more data follows. In such a case
we can directly strdup the remainder of the string.

Ref: https://github.com/linuxwacom/libwacom/issues/85
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>