meson: make building tests optional
Add the option -Dtests=false for environments where we don't care about the
tests. And in turn make the dependencies that are required for the tests
non-optional - provided tests are enabled. You don't get to select a subset of
tests to run because that doesn't help anyone.

Except the test for the right libtool version, that one is never optional.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>