data: explicitly use ls -v to check for files
ls allows to specify a sort order but not for the default alphabetical order.
The result is that if ls has system-specific defaults our test fails because
the default order differs from the one we expected when we wrote the
meson.build file. Work around this by forcing versionsort for ls, even though
this means we have to reshuffle the meson.build file.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>