Add multiple AES stylus definitions
Adds definitions for many different AES styli. The precise information
about supported features is hard to come by, but in general most pens
have two buttons (one of which is treated as an eraser) and can only
report pressure. A few exceptions are known or suspected and their
data has been set accordingly.

Most of the new entries have comments that list several pens for a single
ID. The individual tools could potentially be distinguished with the aid
of vendor ID information available in the HID reports, but this data is
not forwarded to userspace by the kernel.

For each pen I have included additional data in the comments that may or
may not be useful. The format of the comments are as follows:

  OEM1 ; BT device VID | BT device PID [ | FLAGS] ("Model Name" / ModelNumber)
  OEM2 ; BT device VID | BT device PID [ | FLAGS] ("Model Name" / ModelNumber)
  etc.

For tools with a Bluetooth button, the VID:PID of the emulated keyboard
device is given. Tools without such a button use placeholder values of
VID_NONE and 0x0000. Additional flags provide further information:

  * BAT_SWAP / BAT_CHRG: Whether the pen uses replaceable batteries or
    has an internal rechargeable battery.

  * BAT_HID / BAT_GATT: Whether the battery level for the Bluetooth button
    is available through HID or through GATT.

  * BAT_SHARED: Indicates the pen has a single battery powering both the
    AES and Bluetooth portions of the pen.

  * LONGPRESS: Bluetooth button supports the long-press gesture.

  * PROX: Proximity sensing feature available for Bluetooth button.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>