Introduce WSTYLUS_MOBILE type for AES devices
AES styli are distinct enough from the existing stylus classes to
deserve one of their own. This class is named WSTYLUS_MOBILE both to
reflect the fact that AES sensors are only found on tablet PCs and to
allow the potential inclusion of other similar sensors which may not
use AES tech.

Stylus test cases are added to cover the future addition of AES styli
to libwacom.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>