Replace "IsEraser" concept with more fine-grained "EraserType"
Makes it possible for applications to understand where an eraser may
be found on the pen. This could be used e.g. to decide between showing an
image of a pen with an eraser on the back, no eraser, or (the new case)
an additional button that control erasing.

The "is_eraser" function is still kept, but now acts as a convenience
method for identifying erasers of any type.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>