data: add a test for receiver IDs
We don't want any .tablet entries with the receiver ID as that may match
multiple devices. So let's add a check for it, conveniently integrated via
pytest so all we need is invoke pytest and it'll search recursively for tests
to execute.

The actual test searches the directory for .tablet files and runs each through
a config parser to extract the matches.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>