test: Do not allow tablet definition files to match wireless dongle
Several tablets use a special USB dongle to enable wireless use. This dongle
has a VID:PID of 056a:0084 and is not itself considered a tablet. Tablets
which use the dongle should expose their own VID:PID through the dongle.
This commit adds a test to make sure that nobody accidentally tries to
add the dongle as a match for any tablet.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>