tools: switch from a giant udev rulest to using hwdb entries
More efficient and easier to understand and read too. Plus, this set works
with uinput devices out of the box without the need for extra rules.

The new approach is split into a ruleset which now only tags the old serial
devices. Those two lines are easier to keep here than device a hwdb lookup
scheme.

The ruleset imports from the hwdb builtin with the device's modalias, see
/sys/class/device/input/eventXXX/device/modalias. This provides the
bus/vendor/product lookup. And for the prefix use libwacom with the device
name which allows our Finger and Pad matches.

Since we cannot unset properties through a hwdb lookup, we have an extra rule
to set all the joystick 0 envs to the empty string.

The hwdb file itself is generated, note that you must to run systemd-hwdb
update whenever this file changes.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>