data: add GAOMON S56K tablet definition
This tablet masquerades as a Huion (vendor 256c) tablet despite not
being the same manufacturer.

The pen is visually identical to that from the Huion 420, so the Styli
parameter has been copied from it.

The exposed device names are the same as in the Huion 420, with "Huion"
replaced by "Tablet Monitor".