travis: Install Ninja from pip
Meson errors out with a "Could not detect Ninja v1.5 or newer" message
when run on Travis. We already install Ninja 1.5.1, so you'd think there
wouldn't be a problem, but the error message itself is misleading. It
turns out that Meson 0.54.0 (which is what we're currently pulling from
pip) actually depends on Ninja 1.7. Since this version is not available
from the Ubuntu Xenial repositories, we swap over to installing Ninja
from pip as well.

Link: https://mesonbuild.com/Release-notes-for-0-54-0.html#ninja-version-requirement-bumped-to-17
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>