Replace "HasEraser" concept with "PairedStylusIds"
Makes it possible to explicitly link two or more IDs together into a
single pen. The test suite immediately benefits from being able to
rigorously check that pens which should "have" an eraser "do" in fact
have one.

The "has_eraser" function is still kept, but now acts as a convenience
method for identifying tools which are paried with an eraser of any type.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>