test: Add test to check that erasers have corresponding styli
We currently check to see if styli have erasers, but we're missing the
corresponding check to ensure erasers have styli. Lets add that in. We
expect EMR erasers to be linked to a seperate stylus, while we expect
AES erasers to not be linked to anything (since their erasers are just
buttons).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>