Deprecate the WacomClass in the API
To my knowledge, they are not used by anyone outside libwacom itself (which
uses it to set up default button mappings where EvdevCodes is missing) and
their usefulness has degraded over time anyway (any post-I5 device is now also
Intuos5) and the vast majority of devices now are ISDV4. So for any device
bought in the last few years, it's either Cintiq, Intuos5 or ISDV5.

And non-Wacom tablets are even harder to categorize.

Let's deprecate this so if there are any users, they get a compiler warning
about it at least.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>