meson.build: ignore Makefile.in when installing data
A tarball generated with autotools has a generated Makefile.in in every
directory we have a Makefile.am. Using install_subdir() we end up installing
that file unless we explicitly exclude it. This doesn't happen when building
from git (unless the git tree is dirty from a previous autotools setup).

And sneak a "space before tab" whitespace change in while we're in that
vincinity.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>