Drop the show-svg-image tool
This tool uses GTK2 and needs updates. It doesn't provide a huge amount of
benefit. For testing whether an image works, users can drop the new svg file
into the system directories and trigger the GNOME overlay.
So instead of spending time and effort updating this tool, let's just drop it
to get rid of the old dependencies.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>