Add GitHub Actions with the goal of replacing Travis
Travis CI is about to severely limit the free tiers. Specifically, we would
get 1000 minutes (which at current use would last us a few months) but once
they're up, we have to request more on a case-by-case basis.
This is not workable.

Switch to GitHub actions instead. This patch convers the autogen, meson and
build-from-tarball parts over.

It does **not** provide:
- ppc64le
- FreeBSD
- coverity
In the ppc64le and FreeBSD case I don't know how how to get those working.
There doesn't seem to be any available runners for either of those.

Coverity is just a matter of porting from travis to the github action syntax.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>