database: switch file loading to use opendir/readdir
Makes cleanup much easier and we're only going through the files one-by-one
anyway, there is never a need to have the list of entries available.

Same for the dbverify test. We leave the scandir usage in
libwacom-list-local-devices because it provides us with the convenient sort
mechanism that we'd otherwise need to do ourselves.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>