Support loading tablet/stylus files from /etc/libwacom/
To add a new .tablet file, a user would have to add that file to /usr/share.
This location may be read-only on some systems, and any file there will
conflict with files provided by the distribution package. In the best case,
the user's file will be overwritten on update, in the worst case a file name
change leaves both files in place, causing conflicts.

The traditional approach is to have distribution-provided files in /usr and
host-specific packages in /etc, so let's do the same here: add /etc/libwacom
as default lookup path for .tablet and .stylus files.

This requires a change in the file loading: where a directory does not exist
we count it as a non-failure and continue.

In the CI, we have a set of jobs that split the database in various ways
across /usr and /etc. Then we compare the device list before/after the split
and expect them to be the same.

Fixes #327

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>