tools: add a libwacom-show-stylus tool to easily check for stylus IDs
This is an interactive tool to allow users to easily find their stylus IDs.
libinput debug-tablet provides something similar but it's useful to have this
as part of our own repository.

After ctrl+c it prints a suggested Styli line with the set of groups that
belong to this tablet. This is the quick-and-dirty version, probably good
enough for most cases.

Example output from running without arguments:

    $ sudo libwacom-show-stylus
    Using "Wacom Intuos Pro M Pen": /dev/input/event23
    Tool id 0x842 serial 0x748012bf in-proximity: False
    Tool id 0x84a serial 0x748012bf in-proximity: False
    ^CTerminating
    Suggested line for .tablet file:
    Styli=mobilestudio

Nice side-effect: where a prefix is configured but does not (yet) have files
in it, we fall back to the default /usr/share. Too niche to worry about, imo.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>