tools: install our list-devices helper tool
Build it twice, once as list-devices, once as libwacom-list-devices. The
former uses the source tree's data files, the latter the installed data files.

Having this as installed tools makes it easier to debug situations where it's
not clear whether the data file is picked up correctly or whether it's not
applied correctly to the device. It also makes it easier to figure out "is
this device supported by my libwacom version".

Fixes #367

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>