tools: replace the hwdb generator with a python script
This script produces the same output, albeit the ordering is now grouped by
bustype (in alphabetical order). So all Bluetooth devices come first, then the
USB devices - previously the devices were mixed together based on vid/pid/name.
It also fixes a minor sorting issue with the XP Pen Start G640.

The Python script needs to emulate some of the libwacom database logic to
extract the keys from the .tablet files, but in return we get:
- a script that can be run from the source tree (useful for cross-compiling)
- a script that can be run by a user after dropping a file into /etc/libwacom
  so they don't have the manually write a hwdb entry

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>