libwacom: switch the matches to a GArray
Slightly easier to deal with, especially in regards to resizing and copying
it. Since GArray stores the len as guint, we switch everything that needs it
to an unsigned version too - including the device->match index for the
preferred match. This makes some comparisons obsolete, it was never negative
anyway.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>