tool: change the hwdb generator to the libwacom-update-db tool
Previously, the tool generated the hwdb and the rest was left to the user,
requiring more hand-holding in bug reports and more extensive documentation.

Change this to be a one-stop tool that takes care of all steps required by a
user, so a user can just run this tool with /etc/libwacom as argument and the
hwdb will be updated. Notably: the default invocation with /etc/libwacom an
*auxilary* hwdb based on the user's files rather than writing
out the whole one, overwriting the system one.

This is the 99%-case version of the tool, custom prefixes etc. will not be
handled as well as they can be but that's a niche case anyway and not worth
the effort. The vast majority of users will put the new file in to
/etc/libwacom and have the system installation in /usr/share/libwacom.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>