Switch the buttons to a GHashTable
Slightly simpler to manage, we keep a hashtable around with the alphabetic
button as key and a new struct with flags + code as value. This removes the
need for all the seperate array + length tracking.

Minor regression: a caller passing invalid button codes will simply return the
zero values now, without a glib error message.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>