Add libwacom_database_ref/unref
This enables a client to re-use a database context without having to jump
through hoops. Specifically libinput right now wraps the libwacom database
into refcounter struct for that reason - it's easier if we can just use a
libwacom-provided function for this use-case.

This deprecates libwacom_database_destroy(), still works the same as before
assuming the refcount is correct.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>