Overhaul match generation logic
Make this less nested and easier to understand. For each match string:
- create a new match object
- add that match object to the device
- first time only: set the match to be the default match

libwacom_add_match() is a bit peculiar at the moment: if called with a new
match it merely appends to the list but leaves device->match unmodified (at
zero by default). If called with an existing match string, it'll point
device->match to the match in the list.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>